Add timestamps to mvn logs for build-only measurement
This is work in progress, not suitable for merging to master.

Requests timestamping log lines of Maven by adding the necessary
logger properties to IMPALA_MAVEN_OPTIONS. This is currently done
only for bin/jenkins/build-only.sh, as the primary purpose is
to measure build time variance in the FE Java builds.

Change-Id: If1d0a44f70ee95b6499d0405bc5e5296064edd19