WIP: Add CentOS 8.1 support to bootstrap_system.sh
CentOS 8.1 is a new major version of the CentOS family.
It is now stable and popular enough to start supporting it for Impala
development.

This patch adds the preparation steps for a raw CentOS 8.1 system
to support Impala development and testing, both on a standalone computer
or VM, or inside a Docker container.

Note: do not merge, this is still work in progress.

Change-Id: I67a58ec007219020e1fb562216d7a0d1ff38b0bd