Install more toolchain-specific packages on ARM machines
This is required because ARM builds of the toolchain binaries are not
(yet) present in s3://native-toolchain, so ARM builds have to build the
toolchain locally, before being able to build Impala.

The toolchain build needs extra OS packages compared to a regular Impala
builds, some of which were missing.

Change-Id: I1bf6193027d691d3ded727cb59424c5dc9963ea9