IMPALA-12710: Propagate IMPALA_TOOLCHAIN_HOST for dockerized test runs
The driver script bin/jenkins/dockerized-impala-bootstrap-and-test.sh
uses a two-phase approach to initialize the build environment for Impala.
This approach requires explicit actions to preserve the values of
environment variables initialized before the script runs.

IMPALA_TOOLCHAIN_HOST is one of these variables: it allows pulling the
binary toolchain packages from alternative locations. It was missing
from the list of safeguarded variables, so this patch adds it.

Change-Id: I6b7f60167eeb224e35291803e4a819383ff1e346