IMPALA-13223: Fix bootstrap-build.sh for platforms without Python2
bin/bootstrap-build.sh did not distinguish various version of the Ubuntu
platform, and attempted to install unversioned Python packages
(python-dev and python-setuptools) even on newer versions, which don't
support Python 2 any longer (e.g. Ubuntu 22.04 and 24.04). On older
versions these packages are still useful, so at this point it is not
feasible just to drop them.

This patch makes these packages optional: they are added to the list of
packages to be installed only if they actually exist for the platform.

Tests: ran a compile-only build on Ubuntu 20.04 (still has Python 2) and
on Ubuntu 22.04 (does not support Python 2 any more).

Change-Id: I94ade35395afded4e130b79eab8c27c6171b50d6