PROTOTYPE: Extend Docker container build to hardened base containers
Downstream system vendors, users and customers have lately expressed
interest in consuming Impala in containerized forms, taking advantage of
various specialized, hardened container base image offerings, like
containers based on Wolfi OS (by Chainguard).

This patch enables Impala container images to be built on top of
Wolfi-based containers. It follows a hybrid approach: instead of
replicating the complete Impala build process inside a Wolfi container,
it relies on the Red Hat 9 build of Impala to supply the Impala
binaries. This is made possible by the fact that major library
dependencies of Impala have the same version on Wolfi OS and Red Hat 9,
so binaries built on Red Hat 9 can be run on Wolfi unchanged.

The selection of a hardened base image is controlled by two environment
variables:
IMPALA_HARDENED_DOCKER_BASE: specifies the URI of the base image
IMPALA_USE_HARDENED_BASE_IMAGE: if set to 'true', triggers the use of
the hardened image
Add a new script to install Bash in the Docker image if it's missing,
other scripts require it. Add this to the Docker build context.
Check first for ChainGuard base images (Alpine-based): they lack some of the
files/directories that are assumed existing on other distros/base images.

Change-Id: I73d19dab9238ccdfde6788e618b3058c135b83b5