EXPERIMENT: Install curl-minimal into the env for Dockerised tests
Red Hat 8 and 9 releases preinstall curl-minimal for the package
manager. Unfortunately this usually conflicts with the full-blown curl
package later, as their versions is slightly different.

Curl is included in the debugging toolset that is installed into the
Docker image for Dockerised tests. This patch detects the base OS for
the container image, and installs curl-minimal instead of curl when Red
Hat 8 or 9 are detected so that the above conflict is avoided.

Change-Id: I30fa0f13a77ef2a939a1b754014a78c171443c71