IMPALA-13826: Migrate from imp to importlib in the config generator
Python has deprecated the 'imp' package in Python 3.4, and removed it in
Python 3.12. The deprecation has also started throwing warnings in
versions before 3.12.

The template generator used a single call to imp.load_source to load the
template Python file. This is now replaced with code snippet published
in Python's official documentation.

Change-Id: I472d093eeaac97a380d444a1756b54f825b2d031