Suppress -Wignored-attributes in webserver-test.cc for Ubuntu 24.04
Ubuntu 24.04 upgraded glibc to a version that includes attributes on a
lot of FILE-oriented function prototypes. These attributes can trigger
-Wignored-attributes when using such prototypes in templates, which
breaks the build, as all warnings are treated as errors.

This adds -Wno-ignored-attributes to src/util/webserver-test.cc in a
targeted way to suppress this warning for this test file, to make it
possible to compile it on Ubuntu 24.04 (or any other platform sporting
such a new glibc version).

Change-Id: I1c1cf1e6b2d625edc7f10a33fb6d5394d662a0d4